+++ frontend/src/views/GeneratePage.vue
@@ -1,22 +1,7 @@
 <script setup>
-import InputPrompt from '../components/InputPrompt.vue'
+import InputPrompt from '../components/Generate/InputPrompt.vue'
 </script>
 
 <template>
   <InputPrompt></InputPrompt>
 </template>
-
-<style scoped>
-* {
-  margin: 0;
-  padding: 0;
-  font-family: 'Montserrat', sans-serif;
-}
-
-.container {
-  position: relative;
-  width: 100vw;
-  height: 90vh;
-  overflow: hidden;
-}
-</style>
